Isolate views into a sub python package
