+++ vio/vio/pub/vim/vimapi/baseclient.py
@@ -24,7 +24,7 @@ class baseclient(object):
         self._identityclient = None
         self._glanceclient = None
         self._computeClient = None
-		self._cinderclient = None
+        self._cinderclient = None
 
     def identity(self, data):
         '''Construct compute client based on object.
@@ -51,8 +51,8 @@ class baseclient(object):
 
         self._computeClient = driver_base.VimDriver().compute(data)
         return self._computeClient
-		
-	def cinder(self, data):
+
+    def cinder(self, data):
         if self._cinderclient is not None:
             return self._cinderclient
